Some changes

+++ cvect.h
@@ -65,8 +65,8 @@ int *cross(int *x, int *y)
 
 // FUNCTION TO CALCULATE THE MODULUS OF A VECTOR
 double modVector(int *x)
-{
-    return sqrt((double)(x[0]^2)+(x[1]^2)+(x[2]^2));
+{   
+    return sqrt((x[0]^2)+(x[1]^2)+(x[2]^2));
 }
 /*
 // FUNCTION TO FIND ANGLE BETWEEN TWO VECTORS
